Move profiler code over and correct references.
