+++ .idea/modules/emulators/springcoat-vm/squirreljme.emulators.springcoat-vm.test.iml
@@ -12,6 +12,7 @@
     <orderEntry type="module" module-name="squirreljme.modules.meep-swm.main" />
     <orderEntry type="module" module-name="squirreljme.modules.debug-jdwp.main" />
     <orderEntry type="module" module-name="squirreljme.modules.tool-classfile.main" />
+    <orderEntry type="module" module-name="squirreljme.modules.debug-profiler.main" />
     <orderEntry type="module" module-name="squirreljme.modules.cldc-compact.main" />
     <orderEntry type="module" module-name="squirreljme.modules.zip.main" />
     <orderEntry type="module" module-name="squirreljme.modules.cldc.main" />
